refactor transaction types into its own module
