+++ crates/astria-sequencer/src/lib.rs
@@ -10,3 +10,10 @@ pub mod snapshot;
 pub mod state_ext;
 pub mod telemetry;
 pub mod transaction;
+
+pub(crate) fn hash(s: &[u8]) -> Vec<u8> {
+    use sha2::Digest as _;
+    let mut hasher = sha2::Sha256::new();
+    hasher.update(s);
+    hasher.finalize().to_vec()
+}
